+++ src/DayOfWeekValidator.php
@@ -128,14 +128,14 @@ final class DayOfWeekValidator extends FieldValidator
         return $this->isSatisfied((int) $date->format($format), $fieldExpression);
     }
 
-    public function increment(DateTime|DateTimeImmutable $date, string|null $fieldExpression = null): DateTime|DateTimeImmutable
+    public function increment(DateTimeInterface $date, string|null $fieldExpression = null): DateTimeImmutable
     {
-        return $date->add(new DateInterval('P1D'))->setTime(0, 0);
+        return $this->toDateTimeImmutable($date)->add(new DateInterval('P1D'))->setTime(0, 0);
     }
 
-    public function decrement(DateTime|DateTimeImmutable $date, string|null $fieldExpression = null): DateTime|DateTimeImmutable
+    public function decrement(DateTimeInterface $date, string|null $fieldExpression = null): DateTimeImmutable
     {
-        return $date->sub(new DateInterval('P1D'))->setTime(23, 59);
+        return $this->toDateTimeImmutable($date)->sub(new DateInterval('P1D'))->setTime(23, 59);
     }
 
     public function isValid(string $fieldExpression): bool